Make vmgc collect during allocate
This will attempt to collect during allocate if we don't have enough space.

I'm sure this is buggy.  Right now this works in the basic unit test, but fails to help us recover from an otherwise OOM in safe_wren.

Also removed the Error implementation from GCError (not sure if that's correct?) as part of fixing Displa for GCError to not be self-recursive.

Tried to change all the size values to be more explicit about that theyr'e in bytes.  similarly renamed "space" to new_space" in ObjectVisitor to make clear that it's allocating in the new space (or at least should be).